El contador de datos almacenados se no se almacena, se obtiene al contar las memorias no vacias desde la posicion incial

+++ Core/Src/drivers/include/sim.h
@@ -17,7 +17,7 @@ extern "C" {
  * 
  * BUffer conjunto
 */
-uint8_t sim_buffer_512b(uint8_t*buffer,uint32_t len,uint32_t c);
+uint16_t sim_buffer_512b(uint8_t*buffer,uint32_t len,uint32_t c);
